requirements.txt
nltk
numpy
torch #==1.9.1+cu111 -f https://download.pytorch.org/whl/torch_stable.html
torchvision #==0.10.1+cu111 -f https://download.pytorch.org/whl/torch_stable.html
torchtext #==0.10.1